requirement.txt
tensorflow-gpu==2.3.0
tensorflow-addons>=0.10.0
setuptools>=38.5.1
librosa>=0.7.0
soundfile>=0.10.2
matplotlib>=3.1.0
PyYAML>=3.12
tqdm>=4.26.1
h5py>=2.10.0
unidecode>=1.1.1
inflect>=4.1.0
scikit-learn>=0.22.0
pyworld>=0.2.10
numba<=0.48
jamo>=0.4.1
textgrid
click
g2p_en
numpy
pytest-runner
pytest>=3.3.0
hacking>=1.1.0
